Re-style the header
