+++ src/components/Header.jsx
@@ -1,39 +1,59 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
+import React, { useState } from 'react';
+import { Link, useNavigate } from 'react-router-dom';
 import SearchIcon from '../pics/Thumbnails/Search.png';
-import CartIcon from '../pics/Thumbnails/cart.png';
 import logosmall from '../pics/Thumbnails/LogoSmall.png';
+import '../Header.css';
 
-const Header = () => (
-  <header className="mainheader">
-    <div className="leftSideContainer">
-      <img src={logosmall} alt="logo" className="logo" />
-      <button className="signup">
-        <Link to="/signup">Register/Login</Link>
-      </button>
-      <button className="account">
-      <Link to="/account">Account</Link>
-      </button>
-    </div>
-    <div className="rightSideContainer">
-      <div className="search-container">
-        <div className="search-bar">
-          <input type="text" placeholder="Search products..." />
-          <button className="search-button">
-            <img src={SearchIcon} alt="Search" />
-          </button>
+const Header = () => {
+  const [searchInput, setSearchInput] = useState('');
+  const history = useNavigate();
+
+  const handleSearchInputChange = (event) => {
+    setSearchInput(event.target.value);
+  };
+
+  const handleSearch = () => {
+    // Redirect to the All Product page with the search query
+    history.push(`/all-products?search=${searchInput}`);
+  };
+
+  const handleKeyPress = (event) => {
+    // Redirect on pressing Enter key
+    if (event.key === 'Enter') {
+      handleSearch();
+    }
+  };
+
+  return (
+    <header className="mainheader">
+      <div className="leftSideContainer">
+        <img src={logosmall} alt="logo" className="logo" />
+        <Link to="/signup" className="signup">Sign Up/Register</Link>
+        <Link to="/account" className="account">Account</Link>
+      </div>
+      <div className="rightSideContainer">
+        <div className="search-container">
+          <div className="search-bar">
+            <input
+              type="text"
+              placeholder="Search products..."
+              value={searchInput}
+              onChange={handleSearchInputChange}
+              onKeyPress={handleKeyPress}
+            />
+            <button className="search-button" onClick={handleSearch}>
+              <img src={SearchIcon} alt="Search" />
+            </button>
+          </div>
+        </div>
+        <div className="cart-container">
+          <Link to="/cart" className="cart-button">
+            <span className="cart-text">Cart</span>
+          </Link>
         </div>
       </div>
-      <div className="cart-container">
-        <Link to="/cart">
-          <button className="cart-button">
-            <img src={CartIcon} alt="Cart" />
-            <span className="cart-text">Cart</span>
-          </button>
-        </Link>
-      </div>
-    </div>
-  </header>
-);
+    </header>
+  );
+};
 
 export default Header;
