+++ src/components/Navigation.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
+import '../App.css';
 
 const Navigation = () => (
   <header className="secondSmallerHeader">
